requirements.txt
# ANN-Based Multi-Modal Ischemic Stroke Severity & Progression Prediction
# MSc Biotechnology / Bioinformatics Project
# Python >= 3.9 required

# Core deep learning
tensorflow==2.15.0
keras==2.15.0

# Scientific computing
numpy==1.26.4
scipy==1.12.0
pandas==2.2.1

# Medical imaging
nibabel==5.2.0          # NIfTI MRI file I/O (ISLES/ATLAS format)
SimpleITK==2.3.1        # Advanced MRI processing & registration
pydicom==2.4.4          # DICOM support (if needed)

# Data visualization
matplotlib==3.8.3
seaborn==0.13.2

# Machine learning & evaluation
scikit-learn==1.3.2
shap==0.42.1

# Web deployment
flask==3.0.0
flask-cors==4.0.0
streamlit==1.28.1

# Utilities
tqdm==4.66.2
opencv-python==4.9.0.80
Pillow==10.2.0

# Machine learning & evaluation
scikit-learn==1.4.2
imbalanced-learn==0.12.0   # SMOTE for class imbalance

# Explainability
shap==0.45.0
tf-keras-vis==0.8.5        # Grad-CAM on Keras models

# Feature extraction (radiomics)
pyradiomics==3.1.0

# Visualization
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# Hyperparameter tuning
keras-tuner==1.4.6
optuna==3.5.0

# App / dashboard (Phase 6)
streamlit==1.32.0

# Utilities
tqdm==4.66.2
joblib==1.3.2
PyYAML==6.0.1
python-dotenv==1.0.1
